+++ website/backend/nestjs/bind/src/game2.0/dto/GameDto.ts
@@ -2,6 +2,7 @@ import ProfileDto from './ProfileDto';
 import { RacketDto } from './RacketDto';
 import { BallDto } from './BallDto';
 import { WallDto } from './WallDto';
+import Game from 'src/game2.0/Game';
 
 export class GameDto {
 	nbrPlayer: number;
@@ -11,20 +12,40 @@ export class GameDto {
 	walls: Array<WallDto>;
 	rackets: RacketDto[];
 	profiles: ProfileDto[];
-	constructor(nbrPlayer: number, nbrBall: number) {
+	constructor(game: Game) {
 		this.start = false;
-		this.nbrBall = nbrBall;
-		this.nbrPlayer = nbrPlayer;
+		this.nbrBall = game.nbrBall;
+		this.nbrPlayer = game.nbrPlayer;
 		this.balls = [];
 		this.walls = [];
 		this.rackets = [];
 		this.profiles = [];
-	}
-	clone(): GameDto {
-		const rs = new GameDto(this.nbrPlayer, this.nbrBall);
-		rs.start = this.start;
-		rs.balls = this.balls.map((x) => Object.assign({}, x));
-		rs.walls = this.walls.map((x) => Object.assign({}, x));
-		return rs;
+		let i = 0;
+		this.start = game.start;
+		this.nbrBall = game.balls.length;
+		for (i = 0; i < game.balls.length; ++i) {
+			if (!this.balls[i]) this.balls[i] = new BallDto();
+			this.balls[i].x = game.balls[i].x;
+			this.balls[i].y = game.balls[i].y;
+		}
+		i = 0;
+		game.walls.forEach((wall) => {
+			if (!this.walls[i]) this.walls[i] = new WallDto();
+			this.walls[i].x = wall.x;
+			this.walls[i].y = wall.y;
+			this.walls[i].w = wall.width;
+			this.walls[i].h = wall.height;
+			this.walls[i].rotation = wall.angle;
+			++i;
+		});
+		for (const i in this.rackets) {
+			if (!this.rackets[i]) this.rackets[i] = new RacketDto(); 
+			this.rackets[i].x = game.rackets[i].x;
+			this.rackets[i].y = game.rackets[i].y;
+			this.rackets[i].rotation = game.rackets[i].angle;
+			this.rackets[i].h = game.rackets[i].height;
+			this.rackets[i].w = game.rackets[i].width;
+		}
+		this.profiles = this.profiles;
 	}
 }
